+++ block/internal/syncing/syncer.go
@@ -157,7 +157,7 @@ func (s *Syncer) Stop() error {
 	s.cancelP2PWait(0)
 	s.wg.Wait()
 
-	s.daRetriever.StopBackgroundFetcher()
+	s.fiRetriever.StopBackgroundFetcher()
 
 	s.logger.Info().Msg("syncer stopped")
 	return nil
